requirements.txt
scikit-learn
Pillow
wand
openpecha
numpy